A number of changes after the algorithm got cracked.
It can be shown that using the previous algorithm, with the service name and one
password, you can generate N passwords (corresponding to keyphrases of length
N), one of which will be correct (i.e. if you have a five-word pass phrase, the
attacker can generate passwords for lengths 3,4,5,...,N and test them).
If an attacker gets hold of two passwords, using that approach they can find the correct
key length and generate a passphrase that has the same effect as the words
you're using.
Changes:
1. Changed the text to note that this is experimental and people should use at their
   own risk.
2. Added a second algorithm, with following changes comparede to the first (user
   selectable):
   • The keywords (including the service word) are now shuffled differently each time
     a password is generated for a different service. The letter order of the service
     word is used as a guide for shuffling.
   • Instead of adding, after the weighted sum the second word is multiplied and 1
     added,
   • after each combine operation the matrix is shuffled (i.e. sorted by value) before
     taking mod 74.
This is a quick two-hour implementation, code needs to be cleaned up (sorry for the mess).